Update index.html
I don't know how to add an arrow on the button with a line break to put it above the "Back to Top" text with the skills that I currently have. I will have to study and learn more about JavaScript, as well as possibly having to learn more about HTML5 & CSS3. I also will need to learn more about how the three languages interact with each other