Update index.html
Added code for the back-to-top button from the about page and added it to the resume page. I did this so the user would not have to scroll quite as much. I also did this to keep the design guidelines consistent in order keep the design of the website consistent. I did not know if there even are resume webpages on portfolio websites that have a scroll back to the top button, but I am adding one to this page anyway because I want to